+++ src/widget.rs
@@ -151,17 +151,6 @@ where
     }
 }
 
-pub fn bot_icon<'a, M>(
-    style: impl Fn(&Theme) -> selectable_text::Style + 'a,
-) -> Element<'a, M> {
-    selectable_text(String::from("\u{1F916}"))
-        .line_height(LineHeight::Relative(1.0))
-        .font(*font::ICON)
-        .style(style)
-        .size(ICON_SIZE)
-        .into()
-}
-
 pub fn color_dot<'a, M>(color: Color) -> Element<'a, M> {
     selectable_text(String::from("\u{F111}"))
         .line_height(LineHeight::Relative(1.0))